# openapi/altus4.openapi.yaml
openapi: 3.0.3
info:
  title: Altus4 Core API
  version: 0.4.0
  description: |
    Complete API specification for Altus4 - AI-Enhanced MySQL Full-Text Search Engine.

    ## Authentication

    Altus4 uses dual authentication:
    - **JWT Bearer Tokens**: For user management, API key management, database management, and analytics
    - **API Key Bearer Tokens**: For search operations and B2B service integration

    API Key format: `altus4_sk_{env}_{secret}` where env is 'test' or 'live'.
servers:
  - url: http://localhost:3000
    description: Local development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for user authentication
    apiKeyAuth:
      type: http
      scheme: bearer
      description: API key for search operations (format altus4_sk_{env}_{secret})
  schemas:
    ApiResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          description: Indicates if the request was successful
        data:
          description: Response data (varies by endpoint)
        error:
          type: object
          properties:
            code:
              type: string
              description: Error code for programmatic handling
            message:
              type: string
              description: Human-readable error message
            details:
              description: Additional error details
        meta:
          type: object
          properties:
            timestamp:
              type: string
              format: date-time
            requestId:
              type: string
            version:
              type: string
            executionTime:
              type: number
              description: Request execution time in milliseconds
            apiKeyTier:
              type: string
              description: Present on API key authenticated responses
...
    ApiKey:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        keyPrefix:
          type: string
          description: Public prefix of the API key
        environment:
          type: string
          enum: [test, live]
        permissions:
          type: array
          items:
            type: string
            enum: [search, analytics, admin]
        rateLimitTier:
          type: string
          enum: [free, pro, enterprise]
        expiresAt:
          type: string
          format: date-time
          nullable: true
        lastUsed:
          type: string
          format: date-time
          nullable: true
        lastUsedIp:
          type: string
          nullable: true
        usageCount:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    SearchRequest:
      type: object
      required: [query]
      properties:
        query:
          type: string
          minLength: 1
          maxLength: 500
          description: Search query text
        databases:
          type: array
          items:
            type: string
            format: uuid
          description: Database IDs to search in
        tables:
          type: array
          items:
            type: string
          description: Specific table names to search
        columns:
          type: array
          items:
            type: string
          description: Specific column names to search
        searchMode:
          type: string
          enum: [natural, boolean, semantic]
          default: natural
          description: Search mode to use
        limit:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
          description: Maximum number of results
        offset:
          type: integer
          minimum: 0
          default: 0
          description: Number of results to skip
        includeAnalytics:
          type: boolean
          default: false
          description: Include analytics data in response
    SearchResponse:
      type: object
      properties:
        results:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              table:
                type: string
              database:
                type: string
              relevanceScore:
                type: number
              matchedColumns:
                type: array
                items:
                  type: string
              data:
                type: object
                additionalProperties: true
              snippet:
                type: string
              categories:
                type: array
                items:
                  type: string
        categories:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              count:
                type: integer
              confidence:
                type: number
        suggestions:
          type: array
          items:
            type: object
            properties:
              text:
                type: string
              score:
                type: number
              type:
                type: string
                enum: [spelling, semantic, popular]
        trends:
          type: array
          items:
            type: object
            properties:
              period:
                type: string
                enum: [day, week, month, 3months, 6months, year]
              topQueries:
                type: array
                items:
                  type: string
              queryVolume:
                type: number
              avgResponseTime:
                type: number
              popularCategories:
                type: array
                items:
                  type: string
        queryOptimization:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
                enum: [index, query, schema]
              description:
                type: string
              impact:
                type: string
                enum: [low, medium, high]
              sqlSuggestion:
                type: string
        totalCount:
          type: integer
        executionTime:
          type: number
        page:
          type: integer
        limit:
          type: integer
  responses:
    BadRequest:
      description: Bad Request - Validation error
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        example: VALIDATION_ERROR
                      message:
                        type: string
                        example: Invalid request parameters
    Unauthorized:
      description: Unauthorized - Missing or invalid token
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        example: INVALID_API_KEY
                      message:
                        type: string
                        example: Missing or invalid API key
    Forbidden:
      description: Forbidden - Insufficient permissions
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        example: INSUFFICIENT_PERMISSIONS
                      message:
                        type: string
                        example: API key lacks required permissions
...
    RateLimitExceeded:
      description: Too Many Requests - Rate limit exceeded
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        example: RATE_LIMIT_EXCEEDED
                      message:
                        type: string
                        example: Rate limit exceeded. Try again in 45 seconds.
...
paths:
...
  /api/v1/management/setup:
    post:
      tags: [Management]
      summary: Create first API key
      description: Create the initial API key for a new user account
      security:
        - bearerAuth: []
      responses:
        '201':
          description: First API key created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          apiKey:
                            $ref: '#/components/schemas/ApiKey'
                          secretKey:
                            type: string
                            description: Full API key (only shown once)
                          warning:
                            type: string
                            example: This is the only time the full API key will be shown. Please store it securely.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: API key already exists
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      success:
                        type: boolean
                        example: false
                      error:
                        type: object
                        properties:
                          code:
                            type: string
                            example: API_KEY_EXISTS
                          message:
                            type: string
                            example: API key already exists for this user
...
  /api/v1/management/migration-status:
    get:
      tags: [Management]
      summary: Database migration status
      description: Get the current database migration status
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Migration status retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          userId:
                            type: string
                          hasMigrated:
                            type: boolean
                          recommendedAction:
                            type: string
                          documentation:
                            type: string
        '401':
          $ref: '#/components/responses/Unauthorized'

  # API Key Management Endpoints
...
  /api/v1/search:
    post:
      tags: [Search]
      summary: Execute search
      description: Execute a search across specified databases with AI enhancements
      security:
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
      responses:
        '200':
          description: Search executed successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SearchResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
  /api/v1/search/suggestions:
    get:
      tags: [Search]
      summary: Get search suggestions
      description: Get search suggestions based on partial query input
      security:
        - apiKeyAuth: []
      parameters:
        - in: query
          name: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 100
          description: Partial query for suggestions
        - in: query
          name: databases
          schema:
            type: array
            items:
              type: string
              format: uuid
          description: Database IDs to get suggestions from
        - in: query
          name: tables
          schema:
            type: array
            items:
              type: string
          description: Specific tables to get suggestions from
      responses:
        '200':
          description: Suggestions retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          suggestions:
                            type: array
                            items:
                              type: object
                              properties:
                                text:
                                  type: string
                                score:
                                  type: number
                                type:
                                  type: string
                                  enum: [spelling, semantic, popular]
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/search/analyze:
    post:
      tags: [Search]
      summary: Analyze query performance
      description: Analyze a query for performance and get optimization suggestions
      security:
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query, databases]
              properties:
                query:
                  type: string
                  minLength: 1
                databases:
                  type: array
                  minItems: 1
                  items:
                    type: string
                    format: uuid
      responses:
        '200':
          description: Query analysis completed
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          query:
                            type: string
                          recommendations:
                            type: array
                            items:
                              type: string
                          performance:
                            type: object
                          optimization:
                            type: array
                            items:
                              type: object
                              properties:
                                type:
                                  type: string
                                  enum: [index, query, schema]
                                description:
                                  type: string
                                impact:
                                  type: string
                                  enum: [low, medium, high]
                                sqlSuggestion:
                                  type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/search/trends:
    get:
      tags: [Search]
      summary: Get search trends
      description: Get search trends for the authenticated API key user
      security:
        - apiKeyAuth: []
      responses:
        '200':
          description: Search trends retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          type: object
                          properties:
                            period:
                              type: string
                              enum: [day, week, month, 3months, 6months, year]
                            topQueries:
                              type: array
                              items:
                                type: string
                            queryVolume:
                              type: number
                            avgResponseTime:
                              type: number
                            popularCategories:
                              type: array
                              items:
                                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/search/history:
    get:
      tags: [Search]
      summary: Get search history
      description: Get search history for the authenticated API key user
      security:
        - apiKeyAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Maximum number of results
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Number of results to skip
      responses:
        '200':
          description: Search history retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: string
                            query:
                              type: string
                            database:
                              type: string
                            resultCount:
                              type: integer
                            executionTime:
                              type: number
                            timestamp:
                              type: string
                              format: date-time
        '401':
          $ref: '#/components/responses/Unauthorized'

  # Analytics Endpoints
...